requirements.txt
Django==5.0.6
earthengine-api==0.1.384
google-genai==0.2.2
pandas==1.5.3
shapely==2.0.5
pyproj==3.7.1
numpy==1.26.4
python-dotenv==1.0.1
